+++ CHANGELOG.md
@@ -1,17 +1,56 @@
-#  (2024-03-04)
+# 0.4.0 (2024-03-09)
+
+
+### Bug Fixes
+
+* error 500 if i post a pokemon with a missing element ([61f0031](https://github.com/julienhouyet/Symfony-Pokemon-API/commit/61f00318866d98339c4f0e21286219830cab8a90))
 
 
 ### Features
 
-* install API Platform ([a63bf30](https://github.com/julienhouyet/Symfony-Pokemon-API/commit/a63bf309b23b08aeddec0429aa1c2999d5500d93))
-* install Symfony 7.0.4 ([5e51280](https://github.com/julienhouyet/Symfony-Pokemon-API/commit/5e51280a0043d3ca10e1cf3f1666af58d56c7812))
+* automatic token deletion when user is deleted ([75e763f](https://github.com/julienhouyet/Symfony-Pokemon-API/commit/75e763f13bc1df49d60c6aba2db1126256512f96))
+* create user permission test ([7bfb1ff](https://github.com/julienhouyet/Symfony-Pokemon-API/commit/7bfb1ff9d12b978af70322dd6b55eef45ade074f))
+* create UserHashPasswordStateProcessor ([add9a36](https://github.com/julienhouyet/Symfony-Pokemon-API/commit/add9a36dcc3008e6c9eac130a6371827bf12a589))
+* update access permissions for Users, Pokemons & Types ([cc133c9](https://github.com/julienhouyet/Symfony-Pokemon-API/commit/cc133c99b6e6db3ac3179efad77e2d234968bb4c))
+
+
+
+# 0.3.0 (2024-03-08)
+
+### Features
+
+* access token authenticator ([a24d725](https://github.com/julienhouyet/Symfony-Pokemon-API/commit/a24d7258fbb03f6be33f4ccbd93b8a91db65fbd9))
+* add API token generation on user creation ([c5dd630](https://github.com/julienhouyet/Symfony-Pokemon-API/commit/c5dd63001f1be17e173de65e74050772c6b79af9))
+* add fixtures (pokemon, types, users) ([0e772d8](https://github.com/julienhouyet/Symfony-Pokemon-API/commit/0e772d83476dac7b18e6b30621710184b01e87aa))
+* add Makefile for quick docker run ([fc1dcaa](https://github.com/julienhouyet/Symfony-Pokemon-API/commit/fc1dcaa867c704531c07ecce69bcb94675629ecc))
+* add OpenApiFactoryDecorator ([5bc20ba](https://github.com/julienhouyet/Symfony-Pokemon-API/commit/5bc20bad6fbca1e241fd85031ecdc959ad76c8ad))
+* add PostPokemon test ([72acf37](https://github.com/julienhouyet/Symfony-Pokemon-API/commit/72acf370cfcb56e6cdcf413d8849f433effe2141))
+* add Test for Types & Users ([afc2254](https://github.com/julienhouyet/Symfony-Pokemon-API/commit/afc225403170b1c254735b3481008e1dc661553d))
+* add tests for PUT, PATCH and DELETE a pokemon ([f93e6e5](https://github.com/julienhouyet/Symfony-Pokemon-API/commit/f93e6e5bedc9451f88a96499dddddf5badbadbd1))
+* config tokens scopes ([e54cc1f](https://github.com/julienhouyet/Symfony-Pokemon-API/commit/e54cc1f722186adcd0a80f6d22d17c9ba4154275))
+* create ApiToken entity ([a05fa51](https://github.com/julienhouyet/Symfony-Pokemon-API/commit/a05fa5122bff99b36d1a2a2c8e3cc598c8c21519))
+* passing user values to stimulus ([bd64840](https://github.com/julienhouyet/Symfony-Pokemon-API/commit/bd64840991d367bd7432765666183b27b825898c))
 * remove .env file ([72b88f7](https://github.com/julienhouyet/Symfony-Pokemon-API/commit/72b88f7e4d0f5ecf0386db8ecde55c5b2ce6506b))
+* remove scopes temporarily ([cb1fd42](https://github.com/julienhouyet/Symfony-Pokemon-API/commit/cb1fd425525817a05eaa030896db0a72d413477b))
+* remove scopes temporarily ([1148c8c](https://github.com/julienhouyet/Symfony-Pokemon-API/commit/1148c8cd9d3c56c52db51020aa179932676e6893))
+* update access rules for Pokemon resource ([57a317c](https://github.com/julienhouyet/Symfony-Pokemon-API/commit/57a317ce972a853fd5ab5c2232972fe7a5fff7b8))
 
+# 0.2.0 (2024-03-07)
 
+### Features
 
-#  CHANGELOG
+* add .env with fake data ([11cba18](https://github.com/julienhouyet/Symfony-Pokemon-API/commit/11cba184e0b9b220e11d6f9c5398b1979042f50d))
+* add constraints to User entity ([a9c4038](https://github.com/julienhouyet/Symfony-Pokemon-API/commit/a9c4038e7edd7b8ebac384c5ee28ac2dc62ad356))
+* add logout route ([3f4997f](https://github.com/julienhouyet/Symfony-Pokemon-API/commit/3f4997fadeaa3452daf8fd5b264bae39166f0270))
+* add Pokemon, PokemonType & Type entities ([b84a83f](https://github.com/julienhouyet/Symfony-Pokemon-API/commit/b84a83fdb1c91a9f0da35a44b09c1615d623f014))
+* add read relation beetwen pokemon & type ([7abf6f0](https://github.com/julienhouyet/Symfony-Pokemon-API/commit/7abf6f0317ab6bc003a8eed2e7c22d753e66ecf8))
+* add SearchFilter by types name in Pokemon ([b83601b](https://github.com/julienhouyet/Symfony-Pokemon-API/commit/b83601b3424bd19e8a4c83d8366cf095737ac1c9))
+* create LoginForm ([e16e997](https://github.com/julienhouyet/Symfony-Pokemon-API/commit/e16e997fb244f40634c8c5dcd18ecbab2af3c385))
+* create User entity ([c19c960](https://github.com/julienhouyet/Symfony-Pokemon-API/commit/c19c960ce9ad4f0a84a339f3c9f5c3b8d3d9b06e))
+* handling authentification errors ([1867770](https://github.com/julienhouyet/Symfony-Pokemon-API/commit/1867770a2b76df5f80830745c43b0a9a1479f9d6))
+* handling authentification success ([a5fd252](https://github.com/julienhouyet/Symfony-Pokemon-API/commit/a5fd25289aba587802a10842a1a8477d9ce6fbd5))
 
-## v1.0.0 (2024-03-04)
+# 0.1.0 (2024-03-04)
 
 ### Features